moved more code into AlbumList
